requirements.txt
asgiref==3.8.1
babel==2.17.0
Brotli==1.1.0
cairocffi==1.7.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.7.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
django-tenants==3.9.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
fonttools==4.56.0
idna==3.10
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
pillow==11.1.0
psycopg2==2.9.10
pycparser==2.22
pydyf==0.11.0
PyPDF2==3.0.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2024.2
urllib3==2.3.0
weasyprint==64.1
webencodings==0.5.1
xlrd==2.0.2
zopfli==0.2.3.post1
